updating config file

+++ internal/models/user.go
@@ -1,7 +1,7 @@
 package userModel
 
 type User struct {
-	Id       uint   `form:"id" json:"id"`
+	Id       int64  `form:"id" json:"id"`
 	Name     string `form:"name" json:"name"`
 	Email    string `form:"email" json:"email"`
 	Password string `form:"password" json:"password"`
